@@ -32,7 +32,9 @@ const Header = ({ size }: headerProps) => {
       >
          {size === 'desktop' ? (
             <>
-               <h1>City Guide</h1>
+               <Link to="/">
+                  <h1>CITY GUIDE</h1>
+               </Link>
                <nav className={`textButton ${style.menu}`}>
                   <ul>
                      <li>
@@ -41,13 +43,16 @@ const Header = ({ size }: headerProps) => {
                      <li>
                         <a href="#abonnement">Abonnement</a>
                      </li>
+
                      {role === Role.VISITOR ? (
-                        <Link to="/login">Connexion</Link>
+                        <li>
+                           <Link to="/auth/login">Connexion</Link>
+                        </li>
                      ) : null}
                   </ul>
                   {role === Role.VISITOR ? (
                      <button className={`${style.buttonHeader} textButton`}>
-                        <Link to="/register">Nous rejoindre</Link>
+                        <Link to="/auth/register">Nous rejoindre</Link>
                      </button>
                   ) : (
                      <div className={`${style.avatarButton}`} />
